-integrated sidebar with map
