+++ client/src/app/app.component.html
@@ -1,5 +1,54 @@
-<div class="container">
-  <app-header></app-header>
-  <!-- <app-select></app-select> -->
-  <app-map></app-map>
-</div>
+<html lang="en">
+<head>
+      <meta charset="UTF-8">
+      <meta http-equiv="X-UA-Compatible" content="IE=edge">
+      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <title>UF Parking Map</title>
+</head>
+<mat-toolbar color="primary">
+      <button mat-icon-button (click)="sidenav.toggle()" matTooltip="Menu">
+          <mat-icon>menu</mat-icon>
+      </button>
+      <span> UF Parking Map</span>
+  </mat-toolbar>
+  
+  <mat-sidenav-container class="app-container">
+      <mat-sidenav #sidenav mode="side" [class.mat-elevation-z4]="true" class="app-sidenav">
+      
+            <button mat-button  (click)= "openSettings()">
+            <span>Settings</span>
+            </button>
+            <br>
+
+            <button mat-button  (click)= "openAccount()">
+            <span>Account</span>
+            </button>
+            <br>
+
+            <button mat-button (click)= "openSchedule()">
+            <span>Schedules</span>
+            </button>
+            <br>
+
+            <button mat-button  (click)= "openAbout()">
+            <span>About</span>
+            </button>
+            <br>
+              
+      </mat-sidenav>
+      <mat-sidenav-content>
+          <router-outlet>
+              <div class="app-sidenav-content">
+                  <!-- Map content here -->
+                  <div class="container">
+                    <app-header></app-header>
+                    <!-- <app-select></app-select> -->
+                    <app-map></app-map>
+                  </div>
+              </div>
+          </router-outlet>
+      </mat-sidenav-content>
+  </mat-sidenav-container> 
+</html>
+
+<!--<router-outlet></router-outlet> -->
