+++ client/src/app/app.module.ts
@@ -4,8 +4,27 @@ import { LeafletModule } from '@asymmetrik/ngx-leaflet';
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatSelectModule } from '@angular/material/select';
+import { MatDialogModule } from '@angular/material/dialog';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
+import { MatButtonModule } from '@angular/material/button';
+import { MatListModule } from '@angular/material/list';
+import { MatSidenavModule} from '@angular/material/sidenav';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { RouterModule, Routes }   from '@angular/router';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatTooltipModule } from '@angular/material/tooltip';
+
+import { SettingComponent } from './setting.component';
+import { AccountComponent } from './account.component';
+import { ScheduleComponent } from './schedule.component';
+import { AboutComponent } from './about.component';
+
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {CdkAccordionModule} from '@angular/cdk/accordion';
+
 
 import { AppComponent } from './app.component';
 import { MapComponent } from './components/map/map.component';
@@ -21,8 +40,13 @@ import { DecalService } from './services/decal.service';
     MapComponent,
     HeaderComponent,
     SelectComponent,
-    ButtonComponent
+    ButtonComponent,
+    SettingComponent, 
+    AccountComponent, 
+    ScheduleComponent, 
+    AboutComponent
   ],
+
   imports: [
     BrowserModule,
     LeafletModule,
@@ -31,7 +55,20 @@ import { DecalService } from './services/decal.service';
     MatSelectModule,
     FormsModule,
     HttpClientModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    MatDialogModule,
+    MatButtonModule,
+    MatListModule,
+    MatSidenavModule,
+    MatIconModule,
+    MatMenuModule,
+    MatToolbarModule,
+    RouterModule,
+    RouterModule.forRoot([]),
+    MatExpansionModule,
+    MatTooltipModule,
+    MatFormFieldModule,
+    CdkAccordionModule
   ],
   providers: [DecalService],
   bootstrap: [AppComponent],
