+++ client/src/app/components/header/header.component.html
@@ -1,7 +1,7 @@
 <header>
-  <h1>{{ title }}</h1>
-  <app-button 
+  <!-- <h1>{{ title }}</h1> -->
+  <!-- <app-button 
   color="green" 
   (btnClick)="toggleGetDecal()"
-  ></app-button>
+  ></app-button> -->
 </header>
