+++ client/src/app/schedule.component.ts
