+++ client/src/styles.css
@@ -1,6 +1,7 @@
 /* You can add global styles to this file, and also import other style files */
 @import "~leaflet/dist/leaflet.css";
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400&display=swap");
+@import '~bootstrap/dist/css/bootstrap.min.css';
 
 * {
   box-sizing: border-box;
@@ -79,3 +80,65 @@ body {
 
 html, body { height: 100%; }
 body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+
+.about-style mat-dialog-container {
+
+  width: 500px;
+  height: 300px; 
+  overflow: auto;
+
+  border-style: solid !important;
+  border-color: rgb(20, 25, 139) !important; 
+  border: 2px;
+
+  border-radius: 12px !important;
+  text-align: center !important;
+
+  /*maybe create a div element around the text and set the width or min-width to 250 */
+}
+
+.acct-style mat-dialog-container {
+
+  width: 500px;
+  height: 300px; 
+  overflow: auto;
+
+  border-style: solid !important;
+  border-color: rgb(20, 25, 139) !important; 
+  border: 2px;
+
+  border-radius: 12px !important;
+  text-align: center !important;
+}
+
+.mat-mdc-form-field + .mat-mdc-form-field {
+  margin-left: 8px;
+}
+
+.sched-style mat-dialog-container {
+
+  width: 600px;
+  height: 300px; 
+  overflow: auto;
+
+  border-style: solid !important;
+  border-color: rgb(20, 25, 139) !important; 
+  border: 2px;
+
+  border-radius: 12px !important;
+  text-align: center !important;
+}
+
+.s-style mat-dialog-container {
+
+  width: 600px;
+  height: 300px; 
+  overflow: auto;
+
+  border-style: solid !important;
+  border-color: rgb(20, 25, 139) !important; 
+  border: 2px;
+
+  border-radius: 12px !important;
+  text-align: center !important;
+}
